pyproject.toml
[project]
name = "media-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blinker==1.9.0",
    "click==8.2.1",
    "colorama==0.4.6",
    "flask==3.1.1",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf>=1.2.2",
    "greenlet==3.2.3",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy==2.0.41",
    "typing-extensions==4.14.0",
    "werkzeug==3.1.3",
]
